Refactor code for modal
