+++ src/components/content-div.module.css
@@ -1,0 +1,29 @@
+.topicSubhead {
+    font-size: 16px;
+    font-weight: bold;
+    margin-bottom: 1rem;
+}
+
+.topicSubhead.democrat {
+    color: #11719f;
+}
+
+.topicSubhead.republican {
+    color: #a4201d;
+}
+
+.contentList {
+    margin-left: 1rem;
+}
+
+.contentList li {
+    line-height: 1.2rem;
+    margin-bottom: 0.6rem;
+    font-size: 16px;
+}
+
+@media screen and (max-width: 779px) {
+    .contentList li {
+        line-height: 1.5rem;
+    }
+}
