+++ src/components/modal.js
@@ -1,78 +1,12 @@
 import React from "react"
 // import { Link } from "gatsby"
-import ReactMarkdown from 'react-markdown';
+// import ReactMarkdown from 'react-markdown';
 import cardData from "../data/data.json"
 import modalStyles from "./modal.module.css"
+import ModalContent from "./modal-content.js"
 
 const sanitizeString = (string) =>
     string.split(" ").join("_");
-
-function ContentDiv(props) {
-    const contentBullets = props.data.map((bullet, index) => {
-        if (bullet.indexOf("•") > -1) {
-            const subbullets = bullet.split("•");
-            const subbulletsList = subbullets.slice(1).map((subbullet, index) =>
-                <li key={index}><ReactMarkdown source={subbullet} linkTarget="_blank" /></li>
-            );
-
-            return (
-                <li key={index}><ReactMarkdown source={subbullets[0]} linkTarget="_blank" />
-                    <ul>
-                        {subbulletsList}
-                    </ul>
-                </li>
-            )
-        }
-        else {
-            return <li key={index}><ReactMarkdown source={bullet} linkTarget="_blank" /></li>
-        }
-    });
-
-    return (
-        <div>
-            <h4 className={modalStyles.topicSubhead + " " + (props.party === "Democratic" ? modalStyles.democrat : modalStyles.republican)}>{props.topic}</h4>
-            <ul className={modalStyles.contentList}>
-                {contentBullets}
-            </ul>
-        </div>
-    )
-}
-
-function ModalContent(props) {
-    if(props.view === "Overview") {
-        return (
-            <div className="contentContainer"></div>
-        )
-    }
-    else {
-        const data = cardData[props.candidateLastName][props.view];
-
-        const topics = Object.keys(data).sort();
-        const selectedTopicPos = topics.indexOf(props.topic);
-        topics.splice(selectedTopicPos, 1);
-        const remainingTopics = topics.map((topic) =>
-            <ContentDiv
-                key={topic}
-                party={props.party}
-                topic={topic}
-                data={data[topic]}
-            />
-        );
-
-        return (
-            <div className="contentContainer">
-                <ContentDiv
-                    party={props.party}
-                    topic={props.topic}
-                    data={data[props.topic]}
-                />
-                <div className={modalStyles.separator}><span className={modalStyles.separatorLine}></span>Other proposals by {props.view === "Issue areas" ? "issue area" : "type of tax"}<span className={modalStyles.separatorLine}></span></div>
-                {remainingTopics}
-            </div>
-        )
-    }
-
-}
 
 function Modal(props) {
     if(!props.isOpen) {
@@ -122,6 +56,7 @@ function Modal(props) {
                         view={view}
                         topic={topic}
                         party={party}
+                        isPrint="false"
                     />
                 </div>
             </div>
