+++ src/components/modal.module.css
@@ -103,60 +103,9 @@
     color: #11719F;
 }
 
-.topicSubhead {
-    font-size: 16px;
-    font-weight: bold;
-    margin-bottom: 1rem;
-}
-
-.topicSubhead.democrat {
-    color: #11719f;
-}
-
-.topicSubhead.republican {
-    color: #a4201d;
-}
-
-.contentList {
-    margin-left: 1rem;
-}
-
-.contentList li {
-    line-height: 1.2rem;
-    margin-bottom: 0.6rem;
-    font-size: 16px;
-}
-
-.separator {
-    width: 100%;
-    text-align: center;
-    color: #adabac;
-    font-size: 14px;
-    text-transform: uppercase;
-    margin-bottom: 1.2rem;
-}
-
-.separatorLine {
-    display: inline-block;
-    width: 200px;
-    border: 0.5px solid #adabac;
-    margin: 0 40px 4px 40px;
-}
-
 @media screen and (max-width: 1094px) {
     .modal {
         width: 95vw;
         height: 95vh;
-    }
-
-    .separatorLine {
-        width: 20%;
-        margin: 0 30px 4px 30px;
-    }
-}
-
-@media screen and (max-width: 779px) {
-    .contentList li {
-        line-height: 1.5rem;
     }
 }
