+++ src/pages/print.js
@@ -1,9 +1,10 @@
 import React from "react"
 import { useStaticQuery, graphql } from "gatsby"
-import ReactMarkdown from 'react-markdown';
+import ReactMarkdown from 'react-markdown'
 import logo from "../images/tpcLogo.png"
 import cardData from "../data/data.json"
 import printStyles from "./print.module.css"
+import ModalContent from "../components/modal-content.js"
 
 const unsanitizeString = (string) =>
     string.split("_").join(" ");
@@ -100,72 +101,6 @@ function Card(props) {
     )
 }
 
-function ContentDiv(props) {
-    const contentBullets = props.data.map((bullet, index) => {
-        if (bullet.indexOf("•") > -1) {
-            const subbullets = bullet.split("•");
-            const subbulletsList = subbullets.slice(1).map((subbullet, index) =>
-                <li key={index}><ReactMarkdown source={subbullet} linkTarget="_blank" className={printStyles.printLink} /></li>
-            );
-
-            return (
-                <li key={index}><ReactMarkdown source={subbullets[0]} linkTarget="_blank" className={printStyles.printLink} />
-                    <ul>
-                        {subbulletsList}
-                    </ul>
-                </li>
-            )
-        }
-        else {
-            return <li key={index}><ReactMarkdown source={bullet} linkTarget="_blank" className={printStyles.printLink} /></li>
-        }
-    });
-
-    return (
-        <div>
-            <h4 className={printStyles.topicSubhead + " " + (props.party === "Democratic" ? printStyles.democrat : printStyles.republican)}>Proposal on {props.topic.toLowerCase()}</h4>
-            <ul className={printStyles.contentList}>
-                {contentBullets}
-            </ul>
-        </div>
-    )
-}
-
-function ModalContent(props) {
-    if(props.view === "Overview") {
-        return (
-            <div className="contentContainer"></div>
-        )
-    }
-    else {
-        const data = cardData[props.candidateLastName][props.view];
-
-        const topics = Object.keys(data).sort();
-        const selectedTopicPos = topics.indexOf(props.topic);
-        topics.splice(selectedTopicPos, 1);
-        const remainingTopics = topics.map((topic) =>
-            <ContentDiv
-                party={props.party}
-                topic={topic}
-                data={data[topic]}
-            />
-        );
-
-        return (
-            <div className="contentContainer">
-                <ContentDiv
-                    party={props.party}
-                    topic={props.topic}
-                    data={data[props.topic]}
-                />
-                <div className={printStyles.separator}><span className={printStyles.separatorLine}></span>Other proposals by {props.view === "Issue areas" ? "issue area" : "type of tax"}<span className={printStyles.separatorLine}></span></div>
-                {remainingTopics}
-            </div>
-        )
-    }
-
-}
-
 function PrintPage({ location }) {
     const data = useStaticQuery(graphql`
         query PrintSiteTitleQuery {
@@ -220,6 +155,7 @@ function PrintPage({ location }) {
                     view={queryObj.view}
                     topic={queryObj.topic}
                     party={party}
+                    isPrint="true"
                 />
             </>
         }
